+++ src/services/mission.go
@@ -3,11 +3,11 @@ package services
 import "fdms/src/models"
 
 type MissionService interface {
-	Get(id int) (*models.Mission, error)
+	Get(id int64) (*models.Mission, error)
 	GetAll() ([]models.Mission, error)
 	GetAllMissionSummary() ([]models.MissionSummary, error)
 
 	Create(user *models.Mission) (*models.Mission, error)
 	Update(user *models.Mission) error
-	Delete(id int) error
+	Delete(id int64) error
 }
